pyproject.toml
[project]
name = "arbscanner"
version = "0.1.0"
description = "Cross-platform prediction market arbitrage scanner"
requires-python = ">=3.12"
dependencies = [
    "pmxt",
    "sentence-transformers",
    "anthropic",
    "rich",
    "python-dotenv",
]

[project.scripts]
arbscanner = "arbscanner.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arbscanner"]

[tool.ruff]
target-version = "py312"
line-length = 100

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]
